Fix overflowing indices with files > 1gb
Presently, compressing a file with size over 1024 MB causes attempts to index past available buffers in the BT4 matchfinders. This was occuring due to indexing with a 32-bit value generated by bit shifting `-1<<32`, which due to overflow, creates an unexpectedly large value. As a quick fix, replacing all 32-bit variables with 64-bit values in all internal operarations fixed this issue. In the future, it will be best to look into this issue more, restoring usage of 32-bit values where it is both optimal and non-limiting.